feat(multichain): Enable multi-chain balances (#4234)
### Description

For ACT-910. Enables rendering of multi-chain balances in the wallet, if
enabled for the user in Statsig. This PR introduces a host of new
selectors that allow the wallet to access the new token metadata,
indexed by `tokenId`. Where necessary for this PR, our existing
selectors (i.e., with their existing names) have had their
implementations changed to use `tokenId`-based selectors, and the old
implementations have been moved to `WithAddress`-infixed selectors,
whose usage is marked deprecated.

I've also given the same treatment to our `TokenDisplay` component.
Existing usages of `TokenDisplay` have been changed to be
`LegacyTokenDisplay` and deprecated, and `TokenDisplay` now accepts
`tokenId` instead of `address`. While the implementation is pretty
similar, some offline discussion suggested we pull this out into its own
component rather than trying to stuff more functionality into the old
`TokenDisplay` in a backwards-compatible way.

A note on selectors/hooks: from here on out, individual features
throughout the app will potentially want to support different sets of
networks. In order to support this, the "new" versions of selectors are
actually created on-the-fly via higher order functions which accept a
list of `NetworkId`s to filter on. Since this pattern is slightly wonky
(passing paramters to selectors), it is actually _not_ exposed directly
to components. Instead, it is wrapped inside of _hooks_. New components
should use these hooks rather than accessing the higher-order selectors
directly. This is called out in a comment in the
`src/tokens/selectors.ts` file.

### Test plan

Manual and unit tested.

### Related issues

- Fixes ACT-910.

### Backwards compatibility

Yes.