feat(multichain): Use new dynamic config for multichain features (#4265)
### Description

For ACT-931. Replaces three of our existing feature gates with a single
Statsig Dynamic Config (available
[here](https://console.statsig.com/4plizaPmWwPL21ASV4QAO0/dynamic_configs/multi_chain_features)).

The `show_native_tokens` feature gate is a straggler here in terms of
multichain config. This was added as part of a quick and dirty hack to
get non-Celo transfers showing up in the home feed a while back, and
ought to be removed. I'll remove this in a followup PR.

### Test plan

Manual and unit tested.

### Related issues

- Fixes ACT-931

### Backwards compatibility

Yes.