feat: add pending swap transaction to homefeed (#4351)
### Description

As the title. The UI already exists for this, so this PR simply
dispatches the action to add standby tx after the swap has been
submitted.

I added a small tweak in the swap saga to expect the to and from token
to always exist in the redux state, so that we can use the token
decimals etc. in the saga. This should be the case anyway, if the tokens
did not exist in the redux state then the user wouldn't have been able
to select them earlier in the swap flow.

I also noticed that sometimes the `unvalidatedSwapTransaction` returns
token addresses with capital letters, so included a `toLowerCase` so
that the token lookup can be done successfully.

### Test plan



https://github.com/valora-inc/wallet/assets/20150449/cf1eaccb-fc3d-474f-ab80-57aeedad999f





### Related issues

- Fixes RET-894

### Backwards compatibility

Y