feat: more actions token bottom sheet (#4309)
### Description

Adds a more actions token bottom sheet if the more actions button is
present. Moves actions to a hook called `useActions` that takes a token
as a parameter. Moves `onPressCicoAction` to a shared function that is
exported. It appears we will likely need a new feature flag for swaps as
there is currently a coupling of the swap display in the drawer and this
component which seem unrelated.

| iOS | Android |
| --- | --- |
|
![](https://github.com/valora-inc/wallet/assets/26950305/83b2a2c7-2f64-4433-9e22-0323414d5001
"iOS More Actions Bottom Sheet") |
![](https://github.com/valora-inc/wallet/assets/26950305/f193347c-0f04-40d3-9774-e262989716d6
"Android More Actions Bottom Sheet") |

### Test plan
- [x] Tested locally on iOS
- [x] Tested locally on Android
- [x] Unit tests added

### Related issues

- ACT-917

### Backwards compatibility

Yes