feat(balances): Add button to hide balances on home screen (#4413)
### Description

Adds a button to hide users balances on the home screen. When a user
taps the little eye icon next to their total balance, the balance
becomes `XX.XX`, the transaction feed balances are hidden, and the eye
icon turns to a crossed out eye icon. When it is pressed again it goes
back to normal.

### Test plan



https://github.com/valora-inc/wallet/assets/140328381/c2e3ce34-bd8f-432e-92f0-00a244a702dc



### Related issues

N/A

### Backwards compatibility

Yes