fix(send): avoid showing invite modal for contacts that have an address (#4462)
### Description

When a contact is selected and a CPV lookup is done for an address with
that contact's phone number, the recipient object given to the
SendAmount screen as a prop is not filled in with an address, even if an
address is found for them. Instead, their verification status is just
[updated](https://github.com/valora-inc/wallet/blob/9393d6a4bfd2e5425394e60710e275b0ce30f149/src/send/useFetchRecipientVerificationStatus.ts#L34)
to "VERIFIED" and the user is ushered along in the flow, and it is
expected that redux state is used later on to fill in the recipient
address. This breaks an assumption I made in #4415 that verified
recipients will always have addresses.

Note: contact address lookups for the "new" send flow are still broken,
but that is behind a feature flag and incomplete anyway. Since this bug
is on the release branch, we need to get a stopgap out sooner, and we
can follow up to fix the new flow later.

### Test plan

- jest
- tested manually to verify that invite modal does not erroneously
appear when an address exists for that PN

### Related issues

na

### Backwards compatibility

na