test(e2e): wait for TX receipt (#4463)
### Description

Follow up for https://github.com/valora-inc/wallet/pull/4459 since the
failure still
[happened](https://github.com/valora-inc/wallet/actions/runs/6848121582/job/18617673592)
after merging it, but this time it was clear it wasn't because of an
`undefined` TX hash.

Note: this shows there's a slight difference in our implementation when
using ContractKit vs viem.
i.e. with ContractKit we are waiting for the receipt in the wallet, but
not with viem. I think this is fine though.

### Test plan

- Updated e2e test

### Related issues

- N/A

### Backwards compatibility

Yes