fix(logs): logs sometimes cannot be attached (#4633)
### Description

This should fix logs failing to be attached when sent to Zendesk.

Instead of using `react-native-fs` to read the logs in chunk, we now
create a `Blob` which is kept on the native side,
then it's uploaded. 
The advantage is the data doesn't need to be copied in JS memory and we
should avoid the utf8 read error.
So more performant and will work better with large files.

Note: I noticed my log files are rather large (30-40MB!), we should
address this too. Either by compressing or reducing the unnecessary
noise they contain.

### Test plan

- Updated unit tests
- Manually tested

### Related issues

- Fixes RET-948

### Backwards compatibility

Yes