chore: only add a swap approve transaction if the allowance is not enough (#4640)
### Description

Since gas fees are expensive in some networks, it would be beneficial
for users if we could minimise any redundant transactions sent to the
network. Currently for the swap flow, we will always create an approval
transaction for an ERC-20 sell token. There could possibly be a scenario
where a user has approved a larger amount in the past, and the swap
amount in question still fits into the allowance.

This PR adds a check for the spender allowance, before adding the
approval transaction to the list.

### Test plan

Verified that approval transactions are omitted if there is enough
allowance - this is pretty edge case for a user who is swapping on
Valora only since we only approve the necessary amount. I tested by
hacking an approval that is very large, and checking the subsequent swap
does not contain an approval transaction.

I still think there's value here though, e.g.
1. an approval succeeds but the swap transaction fails. the user can
then try again without having to pay for the approval again.
2. a user is swapping a very large amount, the swap transaction succeeds
but due to slippage, some allowance is left over (enough to cover a
subsequent swap that is small)

### Related issues

- Fixes RET-942

### Backwards compatibility

Y

---------

Co-authored-by: Jean Regisser <jean.regisser@gmail.com>