refactor(swap): remove unneeded state + rename variable (#4663)
### Description

`react-async-hook` already keeps the resolved value for us.

Additionally, this fixes an edge case where if `refreshQuote` was called
multiple times, the returned value could be incorrect (out-of-order).
This could happen because `onSuccess` is always called, even if it's not
the `current` promise.

See
https://github.com/slorber/react-async-hook/blob/9351dab27b4e42623b9acdacbf3524ef220035b1/src/index.ts#L263-L275

Also renames `exchangeRate` -> `quote`.

Note: I noticed another edge case, if we get a quote and then get an
error, we'll try to fetch the quote again. Not a big problem but not
entirely expected. I'll fix that separately. It was already present
before this PR.

### Test plan

- Tests pass
- Also tested manually

### Related issues

- Fixes RET-950

### Backwards compatibility

Yes