fix(sends): Fix ERC20 fee currencies for Celo sends (#4697)
### Description

This PR fixes a bug that prevented sends on Celo using ERC20 (non-Celo)
tokens from working correctly.

Basically, we had a check for the `type` field on the prepared
transaction to ensure that it was `cip42` in order to pull the
`feeCurrency` from it to send to the simulate contract method. This
check was _required_ by TS, since if the `type` isn't `cip42`, then the
`feeCurrency` field does not (or isn't supposed to exist). In reality,
we actually _do_ set the `feeCurrency` field without setting the type to
be `cip42`, so we _never_ pulled the `feeCurrency`, and as a result,
sends using non-Celo ERC20s for gas never actually worked.

This is a stopgap solution - once we actually use the prepared
transaction from the send flow _directly_ (i.e., executing the prepared
transaction, rather than pulling all the fields out and reconstructing a
_new_ transaction), we can get rid of this hack entirely.

### Test plan

Manual tested sends on Celo using stablecoins and CELO as feeCurrency,
tested sending ETH on Ethereum.

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>