feat: price history chart display (#4679)
### Description

Uses the new slice and saga created in #4678 to populate a price chart
that is displayed on the token details. If the
`use_price_history_from_blockchain_api` feature gate is enabled. A
sizable amount of chart display logic was copied from
`src/exchange/CeloGoldHistoryChart.tsx` and that existing component was
marked as deprecated and will be replaced by
`src/priceHistory/PriceHistoryChart.tsx` in the future.

### Test plan

- Tested locally on iOS
- Tested locally on Android
- Unit tests added

### Screenshots

| Celo iOS | Eth iOS | Celo Android | Eth Android |
| ----- | ----- | ----- | ----- |
|
![](https://github.com/valora-inc/wallet/assets/26950305/9dea124e-d4f9-4c4a-be11-a3473f4097ad
"Celo iOS") |
![](https://github.com/valora-inc/wallet/assets/26950305/e2e89ca2-c432-409c-9944-29e76095fe63
"Eth iOS") |
![](https://github.com/valora-inc/wallet/assets/26950305/7f4860b2-9d75-48e1-90a9-600b73767b75
"Celo Android") |
![](https://github.com/valora-inc/wallet/assets/26950305/9033f07f-2d68-46e6-bd6c-8d040fa9f472
"Eth Android")

### Load Videos

- There's a can bit a bit of strange on layout rendering that moves the
prices on the x axis if they would be rendered in the margins, but
otherwise the load is WAI.

#### Load on Celo News Page

https://github.com/valora-inc/wallet/assets/26950305/9cad3a2e-286e-4e1c-a554-6c022bc91d93

#### Eth Load and Celo Price Token Details

https://github.com/valora-inc/wallet/assets/26950305/f928a249-ce5d-4ea0-b869-006d88e5214b

### Related issues

- Fixes ACT-99

### Backwards compatibility

Yes