fix: no usd price warning for swaps (#4799)
### Description

This PR fixes a few things:
1. we want to warn the user that they are swapping into an unfavourable
position, so the "no usd price" warning should really be displayed only
for the "to" token. i found it misleading to see this warning when
selecting a token without usd price as a "from" token.
2. the warning copy is tweaked to work better in both the token
selection flow and in the screen itself. we had previously "this token"
in the copy, which was unclear when used on the screen level.
3. prioritise the "no usd price" warning on the screen above the high
price impact warning, because the price impact warning suggests that
swapping a smaller amount may work out better but that's not always true
for tokens that have no usd price. i think it makes more sense to warn
the user about the lack of usd price in this case.

I had missed testing this scenario because I didn't have any tokens that
weren't priced in the test account I was using at the time. I'll double
check with Kayla on these changes and follow up with any changes
separately.

### Test plan


https://github.com/valora-inc/wallet/assets/20150449/0d08a9ea-11a6-4158-8907-553f73b0eca6

### Related issues

n/a

### Backwards compatibility

Y

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>