fix(payments): create and pass a preparedTx to confirmation screen (#4851)
### Description

Bidali and deeplink payments were broken with the new send flow because
a prepared transaction object wasn't passed to the SendConfirmation
screen. This prepares a transaction and passes it on to the
SendConfirmation screen. If a transaction is not possible, it still
navigates to the screen, but the Send button would be disabled (In the
old send flow, if a tx is not possible, Send was still clickable, but
the tx will fail)

### Test plan

Unit tests, manual. Also E2E tests currently failing in #4845 passes
with this fix

Payment deeplink:


https://github.com/valora-inc/wallet/assets/5062591/4774c91c-557a-40f3-916c-4f8c33c87952

Bidali payment:


https://github.com/valora-inc/wallet/assets/5062591/aaeb8b80-471c-4e12-a83c-5aa17710d284

Exceeds balance:


https://github.com/valora-inc/wallet/assets/5062591/4373dae8-5672-4bb8-830d-0b0c22e77fd1


### Related issues

- Fixes ACT-1089

### Backwards compatibility

Yes

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>