refactor(send): prepare a transaction in send confirmation screen (#4980)
### Description

The prepared tx passed to Send Confirmation cannot be directly executed
for tokens that supports comments because it uses a dummy comment to
prepare the tx. This PR updates the send confirmation screen to prepare
a transaction instead of accepting it as a prop for all tokens. This
turned out to be simpler than conditionally preparing a tx (for tokens
that support comments) and also avoids preparing a tx for payment
requests before navigating to the confirmation screen / modal.

### Test plan

Manual, CI



https://github.com/valora-inc/wallet/assets/5062591/5f09bccf-b0c5-4202-a5e7-ecaba601bd7e



### Related issues

- Part of ACT-1049

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>