feat(cab): Add error banner for CAB delete failure (#5000)
### Description

For
[ACT-767](https://linear.app/valora/issue/ACT-767/backup-deletion-loading-and-error-states).

### Test plan

Manual and unit tested. See video below. You can manually trigger this
to appear without code changes by just backing out of PIN entry after
pressing the delete button.



https://github.com/valora-inc/wallet/assets/569401/e70db41a-d4a4-49be-8bf4-ef61f0656bfa


### Related issues

- Fixes
[ACT-767](https://linear.app/valora/issue/ACT-767/backup-deletion-loading-and-error-states).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)