fix(jumpstart): ensure that jumpstart transactions are prepared correctly (#5090)
### Description

We were using an incorrect address for the jumpstart contract previously
so didn't catch these issues. This PR:
- fixes some approval transactions not added to the prepared
transactions. the enter amount screen was passing the user input token
amount to prepare transactions, however it needs to be the token amount
in smallest units. the incorrect units were causing the approval
transaction to never be added because the spend amount would get rounded
to 0 for entered values under 1.
- add dynamic config for the transaction gas estimate, since we cannot
estimate gas for a transaction without approval

### Test plan

No visual changes, manually verified that the transactions created are
using the expected values.

### Related issues

- Related to RET-994

### Backwards compatibility

Y

### Network scalability

Y