chore(jumpstart): support for reclaim from retired jumpstart contracts (#5200)
### Description

This PR adds a `retiredContractAddresses` to the jumpstart dynamic
config.

If we need to change the contract for whatever reason, we should still
allow users to reclaim their funds that were sent to previous versions
of the contract. The only place we need to do this is in the transaction
classification stage - we should mark transactions to _any_ jumpstart
contract appropriately, from there the jumpstart transaction details
screen will allow reclaim from that contract directly.

Of course this logic relies on the abi of the jumpstart contracts
staying the same, since a single abi is baked into the app.

### Test plan

Unit tests

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

Y