chore(points): Use activityId and pointsAmount throughout (#5255)
### Description

Updates the points implementation + redux to use `activityId` and
`pointsAmount` everywhere, as discussed in Slack.

### Test plan

Manual/unit tested.

### Related issues

Part of
[ACT-1032](https://linear.app/valora/issue/RET-1032/[wallet-fake-backend]-points-activity-history).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)