feat(points): add number ticker for balance (#5345)
### Description

As the title. The approach was largely inspired by [a nice man on
youtube](https://www.youtube.com/watch?v=nhshusc5ya0&list=PL7ycojg9_ZoJwnayMF8n-VlhdXnE_grdg&index=1).

I've added support for decimals in case we want to use this for the
wallet balance.

The ticker animation will run any time the value changes, so it will
work automatically for when the points balance changes and the user
remains on the screen.

The downside with animations is that there's a lot of smoke and mirrors
that mess with the dom. For this ticker, we're basically animating
columns of text up and down. From the dom it no longer looks like the
actual balance and we lose the ability to assert on its value from the
unit tests.

### Test plan

Without decimals:


https://github.com/valora-inc/wallet/assets/20150449/a9ce407e-544e-4a69-85c1-9219f629360b


With decimals:


https://github.com/valora-inc/wallet/assets/20150449/4128f00a-0949-4321-98ba-8dda0cd23c89


### Related issues

- Fixes RET-1072

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)