feat(earn): enable gas subsidy for transactions (#5545)
### Description

Based on the feature gate, enable gas subsidy when preparing and
submitting earn transactions. Also dropped tenderly simulation for gas
estimation because we can use syndicate's rpc for this instead of
alchemy and they should succeed even without enough amount for gas.

### Test plan

CI, manually by enabling feature gate (and overriding the rpc url with
syndicate's url as the proxy isn't ready yet) and trying deposit and
withdraw from wallets with no or low ETH and ensured gas is covered.


https://github.com/valora-inc/wallet/assets/5062591/4d4688bb-c89e-4661-b976-8558e1bbd1c0



https://github.com/valora-inc/wallet/assets/5062591/9d63e052-b7a6-4ee1-a955-d3c88056bad0



### Related issues

- Fixes ACT-1194

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)